services/core-ai/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
pydantic>=2.5,<3
pydantic-settings>=2.2,<3
httpx>=0.27,<1
langgraph>=0.0.39
langsmith>=0.1,<1
redis>=5,<6
python-jose>=3.3,<4
structlog>=24.1,<25
